+++ raptor/multilevel/multilevel.hpp
@@ -40,7 +40,7 @@ namespace raptor
                 levels[0]->tmp.resize(Af->n_rows);
 
                 while (levels[last_level]->A->n_rows > max_coarse &&
-                        (max_levels == -1 || levels.size() < max_levels))
+                        (max_levels == -1 || (int) levels.size() < max_levels))
                 {
                     extend_hierarchy(theta);
                     last_level++;
