+++ raptor/multilevel/par_multilevel.hpp
@@ -51,7 +51,7 @@ namespace raptor
 
                 // Add coarse levels to hierarchy 
                 while (levels[last_level]->A->global_num_rows > max_coarse && 
-                        (max_levels == -1 || levels.size() < max_levels))
+                        (max_levels == -1 || (int) levels.size() < max_levels))
                 {
                     extend_hierarchy(theta, num_smooth_steps);
                     last_level++;
